mempool reaping test
